feat: add date / duration picker
