+++ lib/pages/tasks/add_task_modal.dart
@@ -3,7 +3,7 @@ import 'package:app/blocs/tasks/tasks.bloc.dart';
 import 'package:app/components/forms/app_text_form_field.dart';
 import 'package:app/entities/tasks/tasks.entity.dart';
 import 'package:app/i18n/strings.g.dart';
-import 'package:app/components/forms/due_date_picker_modal.dart';
+import 'package:app/components/forms/task_date_picker_modal.dart';
 import 'package:app/utils/constants.dart';
 import 'package:app/utils/exntensions/date_time_extension.dart';
 import 'package:app/utils/shortcuts.dart';
@@ -79,7 +79,7 @@ class _AddTaskModalState extends State<AddTaskModal> {
                               await showModalBottomSheet(
                                   context: context,
                                   isScrollControlled: true,
-                                  builder: (context) => DueDatePickerModal(
+                                  builder: (context) => TaskDatePickerModal(
                                         onDateChanged: (date) {
                                           setState(() {
                                             _dueDate = date;
